Fix Site.replace - no change if nothing changed
With the current implementation of the Site.proto.replace, it fires the
'change' event even when nothing changed in the new sites attributes.
It's happening because a site gets additional props computed from other
attributes. These props are almost never present in the WP.com API
`/me/sites` attributes so the Site.proto.replace would always fire the
'change' event.